// src/schema/types/AutoSjednavacFetchType.ts
export interface AsssistanceRequestType {
  malfunctionAssistance: boolean;
  variantCode: string;
  carReplacementDays: string;
}

export interface AutoSjednavacFetchRequestType {
  beginDate: string;
  endDate: string; // pokud null tak je na dobu neurčitou
  getLimits: boolean; // při true vrací cenu každého limitu
  getApplicableLimitsOnly: boolean; // při tru vrací balíčky jen pro daného uživatele
  getPackages: boolean; // vrací balíčky pro hav
  getFeConfig?: boolean;
  frequencyCode: keyof typeof frequency;
  installmentValue?: number;
  agreementPlace?: string;
  vehicle: Partial<VehicleInfoType>;
  partner: Partial<PartnerInfoType>;
  occupant?: Partial<PartnerInfoType>;
  mtpl: {
    salesDiscount?: number;
    basicLimit?: string;
    selected: boolean;
  };
  assistance: AsssistanceRequestType;
  additionals: {
    glassLimit: string;
    directClaimManagement: boolean;
    injuryDriverLimit: string;
    injuryLimit: string;
    parents: boolean;
    baggageLimit: string;
    premiumWarranty: boolean;
  };
  casco: {
    selected: boolean;
    salesDiscount?: number;
    animalLimit?: string;
    carAccidentLimit?: string;
    foreignTerritories?: boolean;
    gap?: boolean;
    naturalLimit?: string;
    participation?: string;
    theftLimit?: string;
    carCheck?: {
      type?: string;
      agencyId?: string;
      firstName?: string;
      lastName?: string;
      possibleNow?: boolean;
      email?: string;
      phone?: string;
      address?: Partial<Address> | null;
    };
  };
  insured?: Partial<PartnerInfoType>;
  marketingAgreement?: boolean;
  termsAgreement?: boolean;
  termsAgreement2?: boolean;
  policyValidation?: boolean;
  temp: {
    isInternalUser: boolean;
    isAgent: boolean;
  };
  suitabilityRecord?: {
    post: boolean;
    insurers: string[];
    ownTemplate: boolean;
    recommendation: string;
    difference: string;
    request: string;
  };
  dontHaveSpz?: boolean;
}

const enum frequency {
  SINGLE = 'SINGLE',
  ANNUALY = 'ANNUALY',
  SEMIANNUALY = 'SEMIANNUALY',
  QUARTERLY = 'QUARTERLY',
}